feat: add ai advice suggestion and tasks generation

+++ src/schema/application.schema.ts
@@ -1,4 +1,5 @@
 import Application from "../models/application.model";
+import model from "../services/ai/gemini";
 
 export const typeDefs = `#graphql
   scalar Date
@@ -57,6 +58,7 @@ export const typeDefs = `#graphql
     getAllApplication: [Application]
     getApplicationById(_id: ID!): Application
     getTasksGeneratedByAi(_id: ID!): [Task]
+    getAdviceForApplicationByAi(_id: ID!): String
   }
 
   type Mutation {
@@ -64,7 +66,7 @@ export const typeDefs = `#graphql
     deleteApplication(_id: ID!): Application
     updateApplication(_id: ID!, input: ApplicationInput): Application
     addTaskToApplication(applicationId: ID!, task: TaskInput): Application
-    deleteTaskFromApplication(applicationId: ID!, taskId: ID!): Application
+    removeTaskFromApplication(applicationId: ID!, taskId: ID!): Application
     updateTaskInApplication(applicationId: ID!, taskId: ID!, task: TaskInput): Application
   }
 `;
@@ -80,6 +82,44 @@ export const resolvers = {
       const result = await Application.findById(_id);
       if (!result) throw new Error("Application not found");
       return result;
+    },
+    getTasksGeneratedByAi: async (_, { _id }, context) => {
+      const user = await context.authentication();
+      const application = await Application.findById(_id);
+
+      if (!application) throw new Error("Application not found");
+
+      delete user.password;
+      const result = await model.generateContent([
+        "generate tasks the user must take in order to complete the application",
+        JSON.stringify(user),
+        JSON.stringify(application),
+        "provide result in json format for easy parsing where each task contain title, description, due date, and completion status false by default starting from the current date, where the current date is ",
+        new Date().toISOString(),
+      ]);
+      const content = JSON.parse(result.response.text());
+      return content.tasks;
+    },
+
+    getAdviceForApplicationByAi: async (_, { _id }, context) => {
+      const user = await context.authentication();
+      const application = await Application.findById(_id);
+
+      if (!application) throw new Error("Application not found");
+
+      delete user.password;
+
+      const result = await model.generateContent([
+        "provide advice for the user to improve their application",
+        JSON.stringify(user),
+        JSON.stringify(application),
+        "provide result in long text format contain the advice for the user in format of { advice: 'the advice for the user' }",
+        "don't forget to provide the advice in a friendly and helpful manner",
+        "don't mention any negative feedback, only provide positive feedback",
+        "don't mention the json input format in the advice, rather mention the user and application details",
+      ]);
+      const content = JSON.parse(result.response.text());
+      return content.advice;
     },
   },
 
@@ -124,7 +164,7 @@ export const resolvers = {
       return application;
     },
 
-    deleteTaskFromApplication: async (
+    removeTaskFromApplication: async (
       _,
       { applicationId, taskId },
       context,
@@ -134,6 +174,9 @@ export const resolvers = {
       if (!application) throw new Error("Application not found");
 
       const task = application.tasks.id(taskId);
+      if (!task) {
+        throw new Error("Task not found");
+      }
       task.deleteOne();
       await application.save();
